i have update some Ui in several pages

+++ lib/UI/screens/login&SignUP_UI/Login_screen.dart
@@ -55,7 +55,9 @@ class _LoginScreenState extends State<LoginScreen> {
       );
     } else {
       _showErrorSnack(errorMessage);
-      _auth.currentUser!.delete();
+      if (_auth.currentUser != null) {
+        _auth.currentUser!.delete();
+      }
     }
   }
